cancel download button works
cancel button allows stopping the download of a magazine.

modified further HTMLwriterservice and CPCReaderservice with added functionalities:
added to each classes and implementations a public boolean "cancelDl".
at each download attemp, added a verification that cancelDl is false.
if the variable is set to true, an EndServiceException is thrown.

the Async "system" I choosed for Android was the "Service" one.
it allows downloads to continue even when the app is in background.
However, Android doesn't allow easily to stop an ongoing Service.

So  when the user clicks the "cancel button", it calls stopService,
which calls the destructor of the download service, but it's code
continues running until completion.

So in the destructor we set the "cancelDl" variable of the Reader's and
Writer's services to true, and at the next variable verification,
an EndServiceException is thrown.
All the Service code is surrounded by a "try catch" which catches this
exception, and deletes the file before ending.

all theses modifications should work in desktop version too, even if the
class EndServiceException may have to be reimported. (it is currently in
the "Tools.kt" file, and is just composed of this line:

class EndServiceException: Exception()