Use the Login Activity's cookie for Downloadtask
removed login UI from MainActivity: by default the user will be showed
the Login Activity, and the cookie will be saved on AccountManager.

The user can logout on the mainActivity UI.

Before attempting the download, the app verifies that you are logged.

on DownloadTask, before downloading magasine, verifies that the cookie
is valid: calls "canardpc.com/user", and if it redirects to
"canardpc.com/user/login",cancels the download and prompts to login.