first implementation of magasines list
not yet functional.
the principle is, instead of manually inputting the number of the
magasine in the homepage, you will be prompted with a list of all the
magasines, with their cover.

the code is a mess, because instead of getting and parsing the list of
magasines from the website, I attempt to do magic Date calculations to
guess the numbers and their date of release...

the commit also contains 3 cover drawable, "couv380.jpg" is used only for preview in
the IDE, and the others in case of image retreival problem.

next commit will add libs for image downloading, and image blurring.