added Volley Network library for Async donwloads
JSoup is not well suited for Async requests and Network error handling.
Volley will be used to retreive CPC info outside of the parser, and get
pictures (like magasines covers)

modified basic Volley implementations to work with JSoup, and pass on
the error listener when the page is a 404.

added Unit test that tries to get to the CPC number 1234, to verify it
does not exists.

did you know? Canard PC 999 exists in the website's database!
it even has a 10/10 game!