new magasine chooser UI! (not complete)
instead of having to manually input the number of the magasine you want,
you are showed a list of magasines with their cover.

Click listener are not yet implemented, clicking on the mags will not
download them yet!

The MagListbinder and MagListrecycler do lots of strange and messy
calculations to optimize network requests, like calculating the current
number of the magasine, and inferring the release datte of all previous
mag from it.
The covers URL of a mag is retreived by parsing the home pag of this
mag, but the URLs are then stored in an Hashmap in local storage to prevent
 unnecessary network requests in the future.

Added Blur library to blur the covers on the background.
I will soon add an option to disable it, for performance.