bug fixes and small modifications
DownloadTask: moved the downloaded mag to the folder
"/download/CPC-Hors-ligne/magNuber", under the name "mag.html"

If download is interrupted, the mag is deleted, and an "interrupted"
parameter is set to the end broadcast intent.

NetworkUtils: changed ImageRequest to return a bitmap directly instead
of a Drawable.

IAdapter: changed click listener to have access to the parent activity.
This allow to call methods from this activity.

IRecycler: moved IAdapter creation to onActivityCreated method, to pass
the activity to the adapter.
Deleted unused listener.

tmpUtils: added try-catch on readObjectFile method.