changes in MainActivity
modified downloadMag method to be called from outside,
and added check if a download is not already pending.

save and restore dlStarted and numberDl variables at app
closing/opening, so if the user closes te app while a download is
running, download information will be restored when reopening the app.

updated UI code to work with new layout (pushed in next commit)

fixed "download complete" toast showing when the user canceled a download.