Magasine list binder updated, new click listener
updated methods to work with Bitmaps instead of Drawables
when selecting a magasine, if it has not been downloaded yet,
a Dialog will be shown, with the choice of downloading the mag with or without pictures.

when downloading, call storeCoverOnDisk to save the small cover,
and reuse it in the recyclerView, instead of downloading it again.

if the mag has been downloaded, clicking on it opens it with a web browser.